requirements.txt
Flask == 3.0.0
Flask-CORS == 4.0.0
opencv-python == 4.8.1.78
openai == 1.3.4
numpy == 1.26.1
requests == 2.31.0
google-api-core == 2.12.0
google-cloud-documentai == 2.20.2
google-auth == 2.23.4
spacy == 3.7.2
nltk == 3.8.1
fuzzywuzzy == 0.18.0
numpy == 1.26.0